[articleLayout] add carousel and article list
